Initialise new force field implementing a plastic Cosserat model
The idea of the force field is to consider the strain DoFs of the
Cosserat model as a reduecd version of a full strain tensor (for
instance Green-Lagrange). From these strain components, a new stress
state is computed incrementally at each time step, following the radial
return algorithm and a Von-Mises Hill plasticity model with mixed linear
hardening.