Add visualisation method for plasticity
Because of the way deformations are handled in the Cosserat model, we
implement the visualisation method for plasticity in the
DiscreteCosseratMapping, where a baseline visualisation method is
already implemented for Cosserat. Usually, the draw method associated to
plasticity would be found in the (plastic) force field component, but in
this case displaying Cosserat DoFs in the reference world, from the
local coordinates, is complicated (and done with mappings, to the best
of my knowledge).
This solution raises questions :
  - At the moment, the plasticity data is accessed in the mapping with a
SingleLink to a BeamPlasticLawForceField component. A more generic link
could be implemented towards a BeamHookeLawForceField component (from
which BeamPlasticLawFF inherits), but this would require to implement
common methods for visualisation, or cast dynamically depending on a
"m_isPlastic" field.
  - Is such a link the best solution to retrieve plasticity (forcefield
related) data from the mapping ?