Add interactive controller for catheter insertion
The controller allows to apply a rotation on a Cosserat beam, on a given
frame (among the mapped frames). The insertion is also controlled
manually, with a process changing dynamically the frame on which is
applied the control, while the catheter is moving. The purpose of this
functionnality is to mimick the physician's gesture when inserting a
catheter in a vessel.