Add callback for dynamic control
In the Cosserat mapping component, if the discretisation of the Cosserat
beams and/or the mapped frames changes dynamically (typically due to a
controller), the structure of the Jacobian matrix should be updated
accordingly. For instance, the structure of the Jacobian changes if a
frame is moved from one beam element to another.
In order to detect such changes dynamically, we implement a Callback
method in the DiscreteCosseratMapping component, in order to update the
Jacobian if d_curv_abs_section or d_curv_abs_frames change. We
additionaly check the status of d_componentState at the beginning of key
methods of the mapping, to effectively trigger the callback if the
component state is not valid (including if oen of the two mentioned
variables were changed). This way, we make sure that the component is
updated before its methods are used.