Change for loops behaviour (number of beams)
Until now, At least some loops in the BaseCosserat mapping and the
BeamHookeLawForceField component were browsing through the beam elements
using the size of the associated MechanicalObject (rateAngularDeform),
assuming that the number of beam elements and the size of the DoF vector
are the same (which makes sense).
Here, we are considering the possibility that the size of the DoF vector
is higher than the number of beam elements defined in the Cosserat
mapping. The objective, in the long term, is to be able to dynamically
change the discretisation of the Cosserat beam elements (typically to
simulate the insertion and navigation of interventional instruments).

The first step towards this change is to use the number of beam elements
(given for instance by m_BeamLengthVectors in BaseCosserat) instead of
the size of the DoF vector. This is technically breaking, as it changes
the behaviour of the corresponding methods. This commit is experimental
in that regard.