[Plasticity] Make some plasticity variables inherited
Until now, most variables related to plasticity were defined only in
BeamPlasticLawForceField and were not inherited from BeamHookeLawFF. But
implementation of common methods between the two FF, with different
behaviours, requires an additional level of inheritance. Therefore we
switch the definition of the MechanicalState enum type, and some
accessor functions, from the plastic FF to the elastic FF. Difference is
made between the two by a virtual method isPlastic.

NB: I don't know if this is the best implementation scheme to use here.
The question is open.