Add visualisation method for beam segments
Additionnally to the visualisation of segments between each (output)
Rigid Frame of the mapping, we implement the visualisation of the
(input) beam segments. This visualisation relies on a link with a
forcefield, and on new accessor methods. At the moment, only a cylinder
display is implemented, even when the beam cross section is actually
rectangular (in which case the cylinder's diameter is equal to the
smallest section dimension).