[CoaxialModel] Update navigation controller and doc
Main changes:
  - Addition of the Cosserat Rigid frames (output of the mapping) in the
computation of a new discretisation (step 2). This allows to spread the
frames uniformly over the beam elements, at any time during navigation.
  - Implementation of step 3, propagating a new discretisation to the
Cosserat components (mapping inputs and ouputs + force field).
  - Implementation of the motion of the base, following user inputs to
push or pull the instruments.
  - Homogenization of the scene graph, so that each instrument node has
the same structure, and the same component names. Checks are implemented
to raise errors in case of wrong node names, but not yet at the
component level (to be done).
  - Addition of documentation for the controller

NB: a line is also commented in BaseCosserat.inl. The deletion of this
line shoudl not change the behaviour of the Cosserat mapping, except if
some of the output rigid frames are superimposed (which is the case for
the controller). To be confirmed before removal of the line