[CoaxialModel] Update test scene
Contrarily to the InterventionalRadiologyController in the BeamAdapter
plugin, in the current design of the Cosserat coaxial navigation
controller, each instrument has its own specific MechanicalObject. This
means that the 'beam stock' dedicated to each instrument has to take
into account the effective maximum number of beam which can be affected
to the instrument during navigation. In practice, this number depends on
the new beam discretisation computed at each time step, and on the
maximum number of beam elements defined for the other instruments.

The most straightforward way to solve this issue is to determine, in the
simulation scene, the instrument which has the most important 'beam
stock', and add this number to the stock of each instrument (including
this one). This is probably not the most efficient way to handle the
problem, but it should be viable given the usual dimension of the scenes
using Cosserat models to describe endovascular instruments.