[CoaxialModel] Fix bugs in navigation controller
Two main issues are corrected :

- When an instrument is being navigated alone, there is no need to
force the remaining beams in the instrument 'stock' to 0. The reason for
this is that when the instrument is moving backwards, the extra beams
that have to be removed are automatically set to 0 by the last
*curv_abs_input* and *length* values, which are 0 in this case.
However when more than one instrument are simulated, it is possible that
beam elements are added to an instrument which is not currently being
controlled (to make both instrument's discretisation match). In this
case, the situation described above is no longer valid, and extra beams
have to be explicitly removed when an instrument is moving backward. In
this case, both the corresponding *curv_abs_input* and *length* values
are forced to 0. NB: this doesn't apply to frames, which distribution is
specific to each instrument.

- When more than one instrument are simulated, we have to take into
account that the the update of the components is made by iterating over
the entire sets of new beams and frames (common to all instruments). For
each instrument, we therefore have to keep track of the new beam
curvilinear abscissas and new frames which are not affected to the
instrument. This allows to fill the data fields of the instrument's
components with new values, without leaving older values at indices
which correspond to nodes or frames of other instruments.