[CoaxialModel] Update navigation controller and test scene
This commit completes the implementation of the navigation controller,
with :
  - A method to determine which instruments have coaxial segments, and
the corresponding beam elements
  - An update of the components which enforces the constraints
The base test scene is updated accordingly.

NB: in this commit, we rely on the StiffSpringFF component to enforce
constraints between the instruments. This is only viable as long as we
don't introduce any torsion deformation. In future commits, a dedicated
mapping should be implemented, to constrain only the Cosserat DoFs
corresponding to bending.