[CoaxialModel] Add mapping to constrain the rateAngular DoFs
We add a new mapping, inspired by the RigidDistanceMapping, but
instanciated over Vec3 DoFs only. The idea of this mapping is to compute
the difference between two input Vec3 MechanicalObjects, only on desired
entries (for instance y and z, representing bending in the case of
rateAngularDeform).

This implementation is still WIP. At the moment the mapping is
templated, but the template is not strictly necessary as only one
implementation is provided. The question to keep this component
templated should be answered in the future.