[CoaxialModel] Modify coaxial navigation strategy
Until now, the adopted strategy to relate the beam elements which are
common to different instruments was to enforce constraints (through
actual constraints or force fields) on the Cosserat beam DoFs directly
(i.e. in a 'Cosserat strain space', corresponding to the angular rate
component in the SOFA scene).

A different strategy consists in relying on a mapping, which takes as
inputs DoFs corresponding to coaxial instrument components, and gives as
ouput a 'difference' on these DoFs. This difference becomes the quantity
on which we apply the constraint. This constraint can then be propagated
to the actual DoFs through the mapping mechanism.
The first solution implemented here is to duplicate the
DiscreteCosseratMapping on each instrument, to represent a second set of
Rigid frames positioned only at the extremities of coaxial beam
segments. We then add a RigidDistanceMapping for each instrument pair,
computing the difference between the corresponding 'coaxial frames'.
In this commit, we only implement the tracking of the coaxial frames.
The next step is to add a modified version of the
CosseratNeedleSlidingConstraint component, in order to constraint these
frames in the y and z direction only (the x direction being kept free to
allow sliding).

NB: an alternative to the scene structure described above would be to
apply the mapping on the angular Rate DoFs, instead of the output
coaxial frames. However, the corresponding mapping would have to include
the RigidBase of the Cosserat model, in order to propagate back a moment
generated by the coaxial constraints. Such a mapping is left for future
developments.