[DevPlasticity] Fix bug in beam index / frame computation
The previous implementation of the attribution of a beam index to each
output frame did not take into account the possibity to define 0-length
beams (which can be useful, typically in the context of instrument
navigation simulation).
The new implementation takes into account this scenario and should not
be breaking (to be confirmed).