[CoaxialModel] Definition on a object for instrument
Preparing the implementation of the dynamic interpolation of beam
quantities ('positions' and 'velocities' of the angularRate DoFs), we
implement an 'Instrument' Python object to gather properties and methods
related to each instrument discretisation and structure.

This commit also contains a minor modification for the
CosseratNavigationController: the renaming of instrumentLastBeamIds to
instrumentLastNodeIds, which is more explicit.