[CoaxialModel] Add callback in RigidDistanceMapping
With the current implementation of the coaxial navigation controller, we
compute a Rigid difference between the extremity frames of each
instrument beams on overlapping segments. We then apply a 'constraint'
on this difference, either using an acutal constraint, or a forcefield
component (equivalent to RestShapeSpringForceField).
As the number of overlapping segments changes during the navigation, the
input indices in the RigidDistanceMapping which computes the Rigid
difference are also modified, implying that the dimension of the output
mechanical object should be modified as well. To enforce this dynamic
change, we implement a callback in RigidDistanceMapping, ensuring that
the dimension of the ouput MechanicalObject is recomputed every time the
number of input indices changes.

|!\ The consequence of this implementation is that a mappinf (in this
case the RigidDistanceMapping) is able to handle topological changes
(which are fundamentally not different from dynamic changes of input
indices). A better implementation could be to add topology containers
and modifiers in the scene, associated to the current MechanicalObjects,
in charge of handling these 'false' topological changes related to the
discretisation. However this would significantly increase the number of
components in the scene (which is already quiet complex). This aspect
should be discussed in the future