[CoaxialModel] Rename test scenes
The previous test scene ([...]_two_instruments_coaxial.py) is renamed to
[...]_two_instruments_coaxial_angularRate_constraint.py. This version
corresponds to a coaxial navigation controller in which constraints
between the instruments overlapping segments are enforced directly on
the angularRate DoFs. This poses a problem, as such constraints should
include the RigidBase of the model (requiring the implementation of
specific mappings).
The 'v2' version, enforcing constraints on the output Rigid frames of
the Cosserat mappings, is renamed as the main version of the test scene.