[CoaxialModel] Extend NeedleSlidingConstraint to Rigid3d
The purpose of this modification is to be able to use the exact same
constraint on Rigid3d DoFs, instead of Vec3. The implementation is
inspired from the closely related BilateralInteractionConstraint
component.
This first implementation is a prototype, given that only constraints on
the y and z directions can be taken into account (similarly to what
already existed in the component). In the future, it could be
interesting to extend the functionality to any DoF of a Rigid3d data (3
DoFs of position, 3 of orientation)