[CoaxialModel] Add key points to indicate changes in curvature
Similarly to what exists in the BeamAdapter plugin, we add the
possibility for each instrument to have non-constant curvatures at rest.
Changes in curvature are described by 'key points', concretely
corresponding to the curvilinear abscissas (in the instrument frame,
i.e. between 0 and the length of the instrument) at which the curvature
changes. These curvilinear abscissas are defined in the 'instrument'
object, and must be provided by the user with two additional
informations :
  - restStrains associated to each key point interval
  - the number of beams used to discretise each interval.
/!\ By default, we consider two key points at 0 and the length of the
instrument. These key points must not be provided by the user. If no
other keyPoint is specified, only one restStrain value and one number of
beams are additionally required.

Finally, we introduce a new test scene, to test this functionnality