[CoaxialModel] Change default parameter
The nbIntermediateConstraintFrames parameter of the controller is
changed by default from 1 to 0. If the user wants intermediate coaxial
frames to further constrain the coaxial segments, they should set the
number of coaxial frames DoFs accordingly. A safety check is added in
the controller init method, and the main test scene is adapted following
the modification.