[CoaxialModel] Fix and clean RigidDistanceMapping
This commit aims at fixing incoherences in the computation of Transform
objects, when the new option of the mapping, to compute the difference
in a local frame, is used.
In order to improve the sustainability of the code, an important block
of comments was also added in the apply method, to detail the chosen
naming convention, and the internal mechanisms when using Transforms.
This commit includes the following changes :
  - Renaming of "parent" and "child" variables referring to the inputs
and output of the RigidDistanceMapping, in order to avoid confusion with
the notion of child and parent frames in the context of Transforms.
  - Renaming of Transform variables : object1_H_global to
global_H_object1, when the Transform allows to pass from object 1 local
coordinates to global coordinates. This allows to write smoothly
'multiplications' of Transforms
  - Removal of commented code .
  - Removal of a couple of unused variables
  - Change coding style of block sections (if, for) to SOFA's.

/!\ At the moment, the computations in applyJ and applyJT methods is
lacking a term of cross product (currently commented). This
approximation should be tolerable in the context of coaxial Cosserat
navigation, but could become an issue in simulations where the distance
between mapped frames is more important. This should be fixed in further
commits.