[CoaxialModel] Add reinit() method for callbacks (quickfix)
This commit implements an artificial reinit() method for the
RigidDistanceMapping component, in order to be able to trigger callbacks
from the Python code. The implemented method only checks the
componentState status, in order to trigger a callback if the component
has been modified.
The proper way to implement this would probably be to add a Python
binding for the callback itself.

NB: we need to be able to trigger callbacks explicitly from the Python
code, as the normal callback process can lead to issues with the
navigation scenes structure. Indeed, when using the
FreeNotionAnimationLoop, the addForce method of the RestShapeSpringsFF
defined on the RigidDistanceMapping output is called before the callback
of the mapping, whenever the mapping has been modified. Because of that,
there is a mismatch between the indices in the RestShapeSpringsFF
(modified directly from the Python code), and the size of the output
MechanicalObject of the mapping (normally updated through the callback,
which is called later). Directly calling the callback from the Python
code fixes the issue.