[CoaxialModel] Add parameter for spring constraints in controller
Two options are natively available in SOFA to enforce 'constraints'
between the coaxial beam segments during navigation: either Lagrangian
constraints (applied for instance with a CosseratNeedleSlidingConstraint
component) or penalty constraints (applied typically with a
RestShapeSpringsFF component).

With this commit, we add an option in the CosseratNavigationController
in order to adapt the controller behaviour if springs are used instead
of constraints.
Typically, when the size of the RigidDistanceMapping output
MechanicalObject changes, a RestShapeSpringsFF component applied on this
MO has to be updated, while a CosseratNeedleSlidingConstraint component
don't have to.
The parameter we added in the controller is a simple boolean (false by
default) executing some specific code when springs are used to enforce
constraints. In this case, the RestShapeSpringsFF component is updated
after the mapping. Additionnaly, we change some variable in the
MechanicalMatrixMapper components associated to the RestShapeSpringsFF,
in order to trigger a callback and update the MMM as well. /!\ This
requires a modification of the MMM code, which is (obviously) not part
of this commit, and will probably never be part of the main SOFA core,
as it is not a suitable way to update the component. As the MMM
component should be removed from SOFA core in a not-so-distant future,
we leave this issue pending at the moment. But we stress out the fact
that the code in this commit will crash without the adequate
modification in the MechanicalMatrixMapper code.