[CoaxialModel] Remove constraints on too close coaxial frames
Until now, we only prevented the controller to apply constraints on
frames which were too close to the RigidBase, and on intermediate
coaxial frames too close to a beam extremities.
With this commit, we implement a generic method to remove constraints on
any coaxial frame which is too close to another. Note that this
'removal' only changes the indices passed to the RigidDistanceMapping
(to compute the difference between Rigids, on which the constraints
apply). The number of simulated coaxial frames doesn't change (which
should prevent undesirable side effects)