[CoaxialModel] Add input parameter in navigation controller
Until now, the minimal distance to apply constraint parameter was
automatically calculated from the increment distance input parameter.
However, if the user defines a high increment distance, this method will
cause constraints to be ignored for even somewhat distant frames.
To fix this issue, we switch minimalDistanceForConstraint to an input
parameter, which the user has to provide as well.

We update all active test scenes accordingly.